placeholder/requirements.txt
django==1.10.4
gunicorn==19.6.0
Pillow=3.4.2
